requirements/uav-base-py37.txt
numpy==1.21.6
scipy==1.7.3
matplotlib==3.5.3
pandas==1.1.5
tqdm==4.66.5
pyyaml==6.0.1
colorlog==6.10.1
packaging>=21,<25
Pillow==9.5.0
pypdf==3.17.4
torch==1.13.1
geatpy==2.7.0
